+++ src/steps/service-usage/index.ts
@@ -1,4 +1,9 @@
-import { IntegrationStep } from '@jupiterone/integration-sdk-core';
+import {
+  IntegrationStep,
+  Entity,
+  createDirectRelationship,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
 import { ServiceUsageClient } from './client';
 import { IntegrationConfig, IntegrationStepContext } from '../../types';
 import { createApiServiceEntity } from './converters';
@@ -6,7 +11,9 @@ import {
   API_SERVICE_ENTITY_TYPE,
   STEP_API_SERVICES,
   API_SERVICE_ENTITY_CLASS,
+  PROJECT_API_SERVICE_RELATIONSHIP_TYPE,
 } from './constants';
+import { STEP_PROJECT, PROJECT_ENTITY_TYPE } from '../resource-manager';
 
 export * from './constants';
 
@@ -18,8 +25,20 @@ export async function fetchApiServices(
     instance: { config },
   } = context;
   const client = new ServiceUsageClient({ config });
+  const projectEntity = await jobState.getData<Entity>(PROJECT_ENTITY_TYPE);
   await client.iterateServices(async (service) => {
-    await jobState.addEntity(createApiServiceEntity(service));
+    const serviceEntity = await jobState.addEntity(
+      createApiServiceEntity(service),
+    );
+    const r = createDirectRelationship({
+      _class: RelationshipClass.HAS,
+      from: projectEntity,
+      to: serviceEntity,
+      properties: {
+        _type: PROJECT_API_SERVICE_RELATIONSHIP_TYPE,
+      },
+    });
+    await jobState.addRelationship(r);
   });
 }
 
@@ -34,7 +53,15 @@ export const serviceUsageSteps: IntegrationStep<IntegrationConfig>[] = [
         _class: API_SERVICE_ENTITY_CLASS,
       },
     ],
-    relationships: [],
+    relationships: [
+      {
+        _type: PROJECT_API_SERVICE_RELATIONSHIP_TYPE,
+        sourceType: PROJECT_ENTITY_TYPE,
+        _class: RelationshipClass.HAS,
+        targetType: API_SERVICE_ENTITY_TYPE,
+      },
+    ],
+    dependsOn: [STEP_PROJECT],
     executionHandler: fetchApiServices,
   },
 ];